requirements.txt
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
python-dotenv==0.10.1
Flask-Migrate==2.3.1
TogglPy==0.1.1
pytest==4.0.2
pytest-flask==0.14.0
